feat: selectorsとusecaseを作成
