+++ native/app/store/selectors/index.ts
@@ -1,1 +1,2 @@
 export * from './auth';
+export * from './user';
