+++ native/app/store/selectors/user.ts
@@ -1,0 +1,6 @@
+import { User } from "../models";
+import { AppState } from '~/store/modules';
+
+export const userSelector = (state: AppState): User.Model => {
+  return state.user;
+};
